+++ src/IO/posix.c
@@ -353,13 +353,13 @@ int FTI_RecoverVarPOSIX(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
     }
 
     if (data->size != data->sizeStored) {
-        snprintf(str, sizeof(str), "Cannot recover %ld bytes to protected "
-        "variable (ID %d) size: %ld", data->sizeStored, data->id, data->size);
+        snprintf(str, sizeof(str), "Cannot recover %d bytes to protected "
+        "variable (ID %d) size: %d", data->sizeStored, data->id, data->size);
         FTI_Print(str, FTI_WARN);
         return FTI_NREC;
     }
 
-    long filePos = data->filePos;
+    int32_t filePos = data->filePos;
     if (fseek(fileposix, filePos, SEEK_SET) == 0) {
         fread(data->ptr, 1, data->size, fileposix);
         if (ferror(fileposix)) {
